requirements.txt
fastapi
uvicorn
ortools
networkx
requests  # for API calls to external services like ORS
pytest
python-dotenv
